Auto register global components
Store split into 2 modules auth, player. Usage replaced in files
