+++ package.json
@@ -17,6 +17,7 @@
     "core-js": "^3.8.3",
     "firebase": "8",
     "howler": "^2.2.3",
+    "lodash": "^4.17.21",
     "postcss": "^7",
     "register-service-worker": "^1.7.2",
     "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.17",
