+++ src/components/AppHeader.vue
@@ -54,7 +54,9 @@ import { mapMutations, mapState } from "vuex";
 export default {
   name: "AppHeader",
   computed: {
-    ...mapState(["userLoggedIn"]),
+    ...mapState({
+      userLoggedIn: (state) => state.auth.userLoggedIn,
+    }),
     currentLocale() {
       return this.$i18n.locale === "sk" ? "Slovak" : "English";
     },
