+++ src/components/AudioPlayer.vue
@@ -58,7 +58,12 @@ export default {
   name: "AudioPlayer",
   computed: {
     ...mapGetters(["playing"]),
-    ...mapState(["seek", "duration", "playerProgress", "currentSong"]),
+    ...mapState({
+      seek: (state) => state.player.seek,
+      duration: (state) => state.player.duration,
+      playerProgress: (state) => state.player.playerProgress,
+      currentSong: (state) => state.player.currentSong,
+    }),
   },
   methods: {
     ...mapActions(["toggleAudio", "updateSeek"]),
