+++ src/includes/_globals.js
@@ -1,0 +1,26 @@
+/* eslint-disable comma-dangle */
+/* eslint-disable linebreak-style */
+import _ from "lodash";
+
+export default {
+  install(app) {
+    const baseComponents = require.context(
+      "../components/base/",
+      false,
+      /[A-Za-z0-9-_,\s]+\.vue$/i
+    );
+
+    baseComponents.keys().forEach((fileName) => {
+      const componentConfig = baseComponents(fileName);
+      const componentName = _.upperFirst(
+        _.camelCase(fileName.replace(/^\.\//, "")).replace(/\.\w+$/, "")
+      );
+
+      // console.log(fileName, componentName);
+      app.component(
+        `Base${componentName}`,
+        componentConfig.default || componentConfig
+      );
+    });
+  },
+};
