+++ src/main.js
@@ -8,7 +8,8 @@ import Icon from "./directives/icon";
 import "./assets/tailwind.css";
 import "./assets/main.css";
 import i18n from "./includes/i18n";
-import './registerServiceWorker';
+import "./registerServiceWorker";
+import GlobalComponents from "./includes/_globals";
 
 let app;
 
@@ -19,6 +20,7 @@ auth.onAuthStateChanged(() => {
     app.use(store);
     app.use(router);
     app.use(VeeValidatePlugin);
+    app.use(GlobalComponents);
     app.directive("icon", Icon);
 
     app.mount("#app");
