+++ src/views/SongView.vue
@@ -116,7 +116,9 @@ export default {
     };
   },
   computed: {
-    ...mapState(["userLoggedIn"]),
+    ...mapState({
+      userLoggedIn: (state) => state.auth.userLoggedIn,
+    }),
     ...mapGetters(["playing"]),
     sortedComments() {
       return this.comments.slice().sort((a, b) => {
